requirements.txt
kiteconnect
python-dotenv
pyotp
requests
pandas
notebook
matplotlib
